+++ drivers/char/Kconfig
@@ -471,60 +471,6 @@ config ADI
 	  and SSM (Silicon Secured Memory).  Intended consumers of this
 	  driver include crash and makedumpfile.
 
-config RUST_EXAMPLE
-	depends on RUST
-	tristate "Rust Example"
-	default n
-	help
-	  This option builds an example module written in Rust.
-
-	  To compile this as a module, choose M here:
-	  the module will be called rust_example.
-
-	  If unsure, say N.
-
-# XXX: Only for GitHub -- do not commit into mainline
-config RUST_EXAMPLE_2
-	depends on RUST
-	tristate "Rust Example 2"
-	default n
-	help
-	  This option builds a second example module written in Rust.
-	  It is intended to test several Rust modules at the same time.
-
-	  To compile this as a module, choose M here:
-	  the module will be called rust_example_2.
-
-	  If unsure, say N.
-
-# XXX: Only for GitHub -- do not commit into mainline
-config RUST_EXAMPLE_3
-	depends on RUST
-	tristate "Rust Example 3"
-	default n
-	help
-	  This option builds a third example module written in Rust.
-	  It is intended to test several Rust modules at the same time.
-
-	  To compile this as a module, choose M here:
-	  the module will be called rust_example_3.
-
-	  If unsure, say N.
-
-# XXX: Only for GitHub -- do not commit into mainline
-config RUST_EXAMPLE_4
-	depends on RUST
-	tristate "Rust Example 4"
-	default n
-	help
-	  This option builds a fourth example module written in Rust.
-	  It is intended to test several Rust modules at the same time.
-
-	  To compile this as a module, choose M here:
-	  the module will be called rust_example_4.
-
-	  If unsure, say N.
-
 endmenu
 
 config RANDOM_TRUST_CPU
