+++ drivers/char/Makefile
@@ -47,8 +47,3 @@ obj-$(CONFIG_PS3_FLASH)		+= ps3flash.o
 obj-$(CONFIG_XILLYBUS)		+= xillybus/
 obj-$(CONFIG_POWERNV_OP_PANEL)	+= powernv-op-panel.o
 obj-$(CONFIG_ADI)		+= adi.o
-
-obj-$(CONFIG_RUST_EXAMPLE)	+= rust_example.o
-obj-$(CONFIG_RUST_EXAMPLE_2)	+= rust_example_2.o
-obj-$(CONFIG_RUST_EXAMPLE_3)	+= rust_example_3.o
-obj-$(CONFIG_RUST_EXAMPLE_4)	+= rust_example_4.o
